+++ Addon/ElvUI_mMediaTag/core/changelog/3.74.lua
@@ -1,0 +1,10 @@
+mMT.Changelog[374] = {
+	DATE = "18.06.2025",
+	IMPORTANT = { "Attention! With the next version 4.0 of mMT there will be some major changes that will reset some settings." },
+	UPDATE = {
+		"[System]: Add Russian Localization, thx to ZamestoTV.",
+		"[System]: Update Portraits API function.",
+		"[System]: Update Game Version.",
+	},
+	FIX = { "[DT-FPS]: Fixed Bug with FPS Datatext in Classic Versions." },
+}